examples/skills-as-resources/python/pyproject.toml
[project]
name = "skills-as-resources-example"
version = "0.1.0"
description = "Minimal reference implementation: Skills as MCP Resources"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]
license = "Apache-2.0"

[project.scripts]
skills-as-resources = "skills_as_resources.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skills_as_resources"]
